Updates to the menu traversal code (choose_menu_item). It is now capable of handling sprite animation, and has more flexible layout options, and so may be ready for general use.  However, this commit uses it only for the "lessons" menu (which doesn't have sprites); a future commit will implement its usage throughout titlescreen.
Some changes are visually apparent:

1. The layout on the lessons screens has slightly changed. Now there
is a guarantee that the menu item buttons will not overlap the arrows
at the bottom right.

2. The lessons menu text does not flicker anymore when changing the
selected entry.

3. When using the mouse, an item is yellow only if the mouse is within
the button (rather than there always being one selected lesson, even
if the mouse is outside the menu area). This gives it the same
behavior present in the main menu (which seems more polished).

When the main menu code is migrated over, this will also fix a visual
glitch:
4. If the sprite is bigger than the menu button, sometimes the
updating is incomplete (carefully watch the "Arcade game" sprite, the
sprite-tux at the console is sometimes not drawn correctly).



git-svn-id: svn://svn.debian.org/svn/tux4kids/tuxmath/trunk@278 47d30e19-490b-0410-9d18-e851f4e465b3