linewrap: Fix declaration of wrapped_lines
bf89376cd41a9db157b3c9c7b6616c427842ff32 introduced a compile error
wherever this header is included:

    t4k_common.h:248:1: error: multiple storage classes in declaration specifiers
      248 | extern static char wrapped_lines[MAX_LINES][MAX_LINEWIDTH]; //!< Global buffer for wrapped lines.
          | ^~~~~~

A variable can't be both extern and static. The variable must be
declared extern in a header file (which is included in, and compiled as
part of, many source files) but defined in exactly one source file in
this library.

(This variable is not used within this library; but is provided for the
convenience of library users and is used by, at least, tuxmath.)

See discussion on
https://github.com/tux4kids/t4kcommon/commit/bf89376cd41a9db157b3c9c7b6616c427842ff32.